refactor ptiers to tiers
