+++ src/aligned_textgrid/aligned_textgrid.py
@@ -9,7 +9,7 @@ from praatio.textgrid import openTextgrid
 from aligned_textgrid.sequences.sequences import SequenceInterval, Top, Bottom
 from aligned_textgrid.points.points import SequencePoint
 from aligned_textgrid.sequences.tiers import SequenceTier, TierGroup
-from aligned_textgrid.points.ptiers import SequencePointTier, PointsGroup
+from aligned_textgrid.points.tiers import SequencePointTier, PointsGroup
 from typing import Type, Sequence, Literal
 import numpy as np
 import warnings
