+++ src/aligned_textgrid/points/tiers.py
